Fix crashing PDF viewer after changing the storage path
Changing the storage path (and recreating PrefService) caused a crash in
the extension system, because the pointer of the old PrefService was
dangling in ExtensionPrefs. Recreating it as well seems to be the nicest
way to update its dangling part.

Task-number: QTBUG-115188
Change-Id: I46321b1737432c7d1cd6ec41c76e7803e69f2011
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 7af3d247a930baa3262654082c441f8a015889b6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>