PdfMultiPageView: don't force TableView to rebuild from scratch on zoom
It's behaving better now without this workaround, which was also causing
TableView to iteratively pool and reuse delegates for all pages, which
was causing them to actually request images for each page. But at least
forceLayout() still seems to be necessary to cope with the change in
cell size: as the docs say, the columnWidthProvider and
rowHeightProvider functions will return different values after zooming,
so TableView needs to be told to re-evaluate the size and position of
each visible row.

Reverts 4a08034ca3b274e0309530bed4630521b8ad91c9

Pick-to: 6.6 6.5
Fixes: QTBUG-114953
Change-Id: Id84e5359239300d1450868219fe30f402d56c3cb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>