Apply WebEngineSettings to guest views
Animated scrolling is always enabled in the PDF viewer and can't be
disabled by the settings of the embedding web view.

The guest view is still not present at the time when we normally apply
settings to the main WebContents, it is only possible after it was
properly attached.

Change-Id: I6b2c99f67ad90bf028b7299280a9a8b0f31e9f65
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 6b46b4503d3672852b4bf590dd276ff4003b26c8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>