Mark the module as free of Q_FOREACH, except where it isn't
The density of Q_FOREACH uses is non-negligible here, too high for
this author, still unfamiliar with this module, to tackle in a short
amount of time. But they're concentrated in just a few TUs, so pick a
different strategy:

Mark the whole module with QT_NO_FOREACH, to prevent new uses from
creeping in, and whitelist the affected TUs by #undef'ing
QT_NO_FOREACH locally, at the top of each file. For TUs that are part
of a larger executable, this requires these files to be compiled
separately, so add them to NO_PCH_SOURCES (which implies
NO_UNITY_BUILD_SOURCES, too). Created QTBUG-115805 to keep track of
this.

Task-number: QTBUG-115805
Change-Id: I7a80b1820972687ccced24d9edd7a8f5b826556f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>