Prevent QWebEngineCertificateError from being exported twice to QML
We need to export the type as a value type, and also export its enums
in a namespace. This resulted in duplicated entries in the qmltypes
file, which confuses Qt Design Studio.

Fix it by introducing a derived class, and using it for exporting the
enums.

Fixes: QTBUG-115365
Change-Id: I5900f4f76fb6357c9262496cc3cc01b276387e46
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
(cherry picked from commit df4121886dfa11cf90e74c72b1121a96462726bf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>