QtPdf: Fix memory leaks
In QPdfBookmarkModelPrivate::appendChildNode it was not necessary to
call FPDF_LoadPage() just to be able to call FPDF_GetPageHeight():
the BookmarkModel has the QPdfDocument which can do that for it,
using FPDF_GetPageSizeByIndex() instead.

7b8832ca2b84d549c9d374550c3c46b3d4d42a38 added
QQuickPdfDocument::carrierFile() returning a pointer that
QQuickPdfDocument continues to own. QQuickPdfDocument::setSource()
took care of deleting the existing one if we open a different
document, but the destructor did not delete it.

Pick-to: 6.2 6.5 6.6
Task-number: QTBUG-114953
Change-Id: Ie8d9cea456cb91d9541deb575dac17db0d367ead
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>