QPdfSearchModel: remove onDocumentPageCountChanged()
This function was introduced in
d1347ebbe99b467799680bb6e3cee81932751df5 as a bugfix for QTBUG-111067.
The patch was also picked to 6.5, which is not really correct, because
we normally guarantee forward BC in patch releases.

Generally, there is no reason to expose this method to the API - we
could just use a lambda internally. Do it now.

Treat the exposed function as internal, and remove it completely
both from the API and the ABI.

Pick-to: 6.6 6.5
Change-Id: I48b93767c341aca58135872f391fc731561b68d4
Reviewed-by: Michal Klocek <michal.klocek@qt.io>