Do not try to delete BrowserAccessibilityInterface twice
BrowserAccessibilityInterface is deleted when removing it from the
QAccessibleInterface cache. It is removed from the cache in the
BrowserAccessibilityQt destructor explicitly.

On exit, the QAccessibleInterface cache might be destructed earlier than
the corresponding BrowserAccessibilityQt and deletes all the cached
interfaces. Reset the interface pointer to nullptr to make sure that the
destructor doesn't try to delete an already deleted interface.

Pick-to: 6.6.0 6.6 6.5 6.5.3
Fixes: QTBUG-115753
Change-Id: Ide02a84d2230145a3441d9ade47eaa14af897477
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>