Fix Qt accessibility cache logging
The accessibility cache logging may crash:
- when logging an "insert" and the underlying Chromium accessibility
  objects are not yet created.
- when logging a "delete" and the underlying Chromium accessibility
  objects are already destructed.

BrowserAccessibilityManager::GetFromId() is not supposed to return
nullptr but it does with the QtWebEngine integration and it happens when
we try to access BrowserAccessibility too early or too late. Add check
to those BrowserAccessibilityInterface methods which use GetFromId().
These methods are called by the Qt logger on creation and destruction of
the BrowserAccessibilityInterface.

The logging can be enabled with the following environment variable:
QT_LOGGING_RULES="qt.accessibility.cache.debug=true"

Pick-to: 6.5 6.6 6.6.0
Change-Id: Ib97f0acb982b72a5602234bd5845749f69689e32
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>