Fix handling of external URLs in PDFs
The code assumed that the guest view web contents would have a delegate
of type WebContentsDelegateQt (as it does in later versions after some
refactoring) and, due to the lack of RTTI, called a method that does not
exist in GuestViewBase. Fix this for this branch by using the top level
web contents' delegate in this case.

Fixes: QTBUG-117453
Change-Id: I9d32f145bf83ab68f8ee83a5fefa81c800896536
Reviewed-by: Szabolcs David <davidsz@inf.u-szeged.hu>