Handle initial NavigationEntries
Chromium introduced the Initial NavigationEntry that meant to
represent the initial empty document at FrameTree creation time.
It is just a placeholder entry put in the list of NavigationEntries
that gets replaced on the next navigation.
A side effect of this is that we have an extra entry in the list before
we load any page.

This change also reverts the history related tests modified by
106 Adaptations that introduced this behavior.

Fixes: QTBUG-117489
Pick-to: 6.5 6.6
Change-Id: I2738591b681082792544bd884f8ecec54ce9d72d
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>