Support auto accepting non-HTML saves when a path was specified
Images and PDFs had to be explicitly accepted to download, even if a
filePath was given as argument of QWebEnginePage::save(). This behavior
was inconsistent with saving HTML pages (and the documentation).

Adjust the corresponding test case with verifying the download result
of an auto-accepted download, and add a new test case to make sure about
downloads are not accepted by default when there is no desired filePath.

Pick-to: 6.6
Task-number: QTBUG-117624
Change-Id: I90f4cfa485da2831327bcb25571a0b69499d9006
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>