Fix typo in "src/CMakeLists.txt"
"src/CMakeLists.txt" defines a custom command with "OUTPUT",
`copyOutput`.

`copyOutput` is supposedly set a few lines above the custom command but,
due to a typo in the `set` statement, it is not actually set when the
command is defined, resulting in a custom command without a "TARGET" or
"OUTPUT" and thus a CMake error, preventing configuration.

To fix the issue, the type in the `set` statement is now fixed so that
`copyOutput` is correctly initialized.

This amends a09f5c4d5ae11ac338f09de044b72b6fc6ac00cc.

Change-Id: Ia86e64aca6c490acc5d4cfade39ca9d3ccf8c5e4
Reviewed-by: Michal Klocek <michal.klocek@qt.io>