Add crc32 setting
Let's assume user sets march to be armv7,
this means we should not compile crc32 simd
part of zlib as will fail.

Chromium overrides this with armv8-a-aes-crc,
however we do not want to do it as it hides the
issue.

Note there is runtime check for cpu features
with zlib but it assumes armv8 architecture.

We could add the test as CRC32 intrinsics are
only defined in arm_acle.h for armv8 target,
however just extract the arm version from
compiler flags for now.

Change-Id: Ia2ed121873e3dca50f895fc6da517d513196520b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>