Fix cursor position during composition
During input composition, the cursor is set to the starting position of
the string being edited and refrain from updating it throughout the
composition process. Once the string is committed, it is set to the last
position of the committed string. The modified code is designed to
exhibit the same behavior as the widget and quickwidget.

Fixes: QTBUG-118746
Pick-to: 6.6 6.5
Change-Id: I9adb7a9bea9493457a1d9d4b32338cdba30920a1
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
(cherry picked from commit c9550388b9ab1d82454d0c59eaaf7e640d83d750)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>