Add PdfSearchModel.count; disable search up/down buttons if empty
A crash was possible if there are no search results and the user
presses the up/down buttons alongside the search field on the toolbar.
That was easily fixed (in a previous patch); but actually those buttons
should be disabled if there are no results to navigate. To do that, we
need a notifying property rather than depending on the invokable
QAIM::rowCount(). Follow the usual pattern for a one-dimensional model:
add a `count` property in the subclass (as in XmlListModel,
FolderListModel, CalendarModel, and a few others).

Task-number: QTBUG-119776
Change-Id: I51c75b58f3a440c1c470794518a572b384376ea7
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>