QQuickPdfPageImage::load(): fall back to base class impl to avoid crash
Plain Image is ok for loading one PDF page at a time, except that the
pdf plugin has to create its own PDF document object if it cannot reuse
an instance that is already open. So PdfPageImage was created just as an
optimization, with the expectation that the document will be given to
the document property. But in case someone doesn't understand and tries
to set only the inherited Image.source property instead, fall back to
QQuickImageBase::load() to avoid crashing. Amends
7b8832ca2b84d549c9d374550c3c46b3d4d42a38

If both the document and source properties are set, prefer the document
(for the sake of performance).

Add test coverage of other combinations of properties, while we're at it.

Pick-to: 6.6 6.5
Fixes: QTBUG-104767
Task-number: QTBUG-77506
Change-Id: I1ee0d0bb2a6c5f399234ddddd969be02e7a6c020
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit b60c00f4adec9ea4b75af0226b04a7125e166ae2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>