Expand helper process stack to 8MB on Windows
Chromium on Windows is built with a linker flag that ensures an
initial stack size of 8MB, whereas a QtWebEngine build never sets one.
The result is that on Windows the helper process only gets the default
1MB stack, which can cause stack overflows in pages with nested
elements (see issue QTBUG-120420). This change makes sure the relevant
linker flag is added when a Windows build is detected, and the stack
size matches the one in Chromium.

Pick-to: 6.7 6.6
Task-number: QTBUG-120420
Change-Id: I42ff397f1b90c4455188b5711d5b3e9c1780db23
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>