Fix high-dpi scaling in PdfScrollablePageView
Amends 85f82185ef5102a2dbe3b2e437b9723e1652ce57

In PdfScrollablePageView.qml, a property like

  property size pagePointSize: document.pagePointSize(pageNavigator.currentPage)

would get evaluated too early (QPdfDocument::pagePointSize() returns a
default-constructed QSize if d->doc is still null). We need to call
pagePointSize() directly in bindings that need it, so that the function
will be called again each time some other binding dependency changes,
to get the size of the current page in the current document. Perhaps
a binding to the result of a function call does not get invalidated
when the function argument changes; or perhaps behavior changed now
that the QML files tend to be compiled rather than runtime-interpreted
resources.

We also need PdfPageImage's width and height to be bound (as they are in
PdfMultiPageView.qml), to avoid automatically scaling up to sourceSize.

So now we get high-dpi page images again in the single-page example.

Task-number: QTBUG-86948
Pick-to: 6.5 6.6 6.7
Change-Id: I44e8df15c18bd0b752dccc899028e7e7900f6ffb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>