Keep page's background color when loading new URLs
Whenever a new URL is loaded, the background color is reset to white,
even if the user has explicitly set the background color to a different
value. Therefore, it needs to be reapplied when a new render widget is
created.

Fixes: QTBUG-112013
Fixes: QTBUG-120926
Pick-to: 6.6
Change-Id: I6663008ee038f30ea6f5df8ab7d8ce7356b152a1
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 85bfaa5274b5d47437d6e0634746f70a6646bd4e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>