CMake cleanup: Move configure checks before features
We have qtwebengine_configure_checks which aim is to detect
issues, which would prevent to have buildable configuration.
Based on those checks we then should later enable features
like for example building with system libs.
Therefore reverse the current order of evaluation.

Note configure should always succeed, the reason for that
is that we build two modules qtpdf and qtwebengine.
It can happen that only one of those modules is supported
only, therefore configure has to succeed on ci to allow
building the supported module.

Change-Id: Ieaa544189fd90b8ab30e654f99b64526dd1040e9
Reviewed-by: Michael Brüning <michael.bruning@qt.io>