Propagate system light/dark mode to webpages
This change makes sure that Chromium is aware of any switches
between light/dark mode in the system theme. This allows webpages
that use the 'prefers-color-scheme' media feature query to see
the system color scheme, and adapt accordingly. An example is the
following page (set theme to 'OS default' in the top right):
https://developer.mozilla.org/en-US/docs/Web

Change-Id: Ifee20d15bbc1681c6ea2ee016f0283a6fa09fc77
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>