Fix Vulkan rendering with Nvidia on Linux
Nvidia driver does not support GPU memory buffer (GMB) properly. The
NativePixmap is only available if the SharedImage is backed by GMB.
OverlayImageRepresentation only provides NativePixmap to access the
texture memory by default.

ExternalVkImageBacking is a fallback for OzoneImageBacking if GMB is
not supported and native Vulkan rendering is used on Linux (see
//gpu/command_buffer/service/shared_image/shared_image_factory.cc).
It uses VkImage backing instead of GMB by default. This VkImage can be
accessed via SkiaImageRepresentation what we already use for
synchronizing texture rendering into the backing buffer.

As a workaround for Nvidia, the backing VkImage is imported into the
Qt's Vulkan context instead of NativePixmap.

Pick-to: 6.7
Change-Id: I94a29488521473291a5ba547abca34a3dba567a2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>