Fix Quick WebEngine middle mouse button scrolling
In Qt Quick, mouse movement will be sent to WebEngine as a QHoverEvent
instead of a QMouseEvent. This change makes sure that when converting
a QHoverEvent to a WebMouseEvent, the screen position is set correctly.
This, in turn, ensures that Chromium will correctly interpret
mouse movement when performing a middle mouse button scroll;
instead of immediately jumping to the top left of the page, it will
now scroll correctly.

Fixes: QTBUG-120131
Change-Id: Ic5cac761974d422bf1f774616e99e5edbf82f4de
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>