Allow checking whether certificate error comes from a page resource
Currently, all certificate errors are handled the same way, regardless
of whether they come from the main frame or from an outside resource.
On the user side, this means that all certificate errors for resources
trigger a pop-up claiming the page is unsafe. Other browsers do not
do this, and instead automatically reject bad certificates for
resources.

This change extends the QWebEngineCertificateError API to allow
querying for whether the error comes from the main frame, or from
a resource. This information is already supplied by Chromium, but we
previously just ignored it. With this, application developers may now
choose to auto-reject resource errors just as Chrome does.

The simple and quick nano browser examples have been updated to
reject certificate errors coming from a page resource.

Task-number: QTBUG-92009
Change-Id: Id38b21f569318d7a83332d4ba69cb27d14b53430
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>