Workaround for setCursor issue with embedded views
QQuickItem::setCursor maps the cursor position to the window
while the Item's position is mapped to the View.
If the View is moved inside its container, the offset is ignored
on the Item. Ergo View.pos != Item.pos breaking hittesting of setCursor.

Fix this by forwarding Item::setCursor calls to the View.
It is safe since the Item shares its coordinate system
with chromium and works as a proxy to the View.

Fixes: QTBUG-111927
Fixes: QTBUG-123889
Fixes: QTBUG-115929
Pick-to: 6.7 6.6
Change-Id: Idee5ba043774952b554874ce654279cb5029ef1d
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 5043f4ddff206d2ffcb4684fb791ec689c2538f0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>