Fix print preview dialog crashes in printme example
The example runs a nested event loop while waiting for the printer
to return a document. During that event loop, a user was previously
free to click on any of the buttons in the print preview dialog,
whoich would mess up the internal state and cause a crash. This change
adds a filter to that event loop to exclude user input events, and
updates the docs to mention the reasoning behind this pattern.

Fixes: QTBUG-122916
Pick-to: 6.7
Change-Id: I4d3d336ffa9bd23e036772bd78f6f1e5870ab732
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>