Ensure that the correct modifier key is send to the chromium
QtWebEngine converts QtKeyEvent to NativeEvent and then forwards
it to Chromium. During the conversion, the value of
keyEvent->nativeModifiers is passed as the modifier, leading to
this issue. Eg: when using the command modifier, nativeModifiers
yielded 264 instead of the expected value NSEventModifierFlagCommand,
which is 1 << 20. Used QAppleKeyMapper::toCocoaModifiers() to get
NSEventModifierFlags from Qt::KeyboardModifiers

Fixes: QTBUG-122970
Pick-to: 6.7
Change-Id: I0c9d8df796187a04475fa944bf1cf4ba2a14109b
Reviewed-by: Michael Brüning <michael.bruning@qt.io>