Implement optional website permission persistence
This change introduces a new API allowing application developers
to choose whether they want to retain the granted/denied status of
permissions between different pages or browsing sessions. The
previous behavior of asking for permission every time is still
optionally available, but the default behavior is to persist
permissions inbetween sessions (except for off-the-record profiles,
where the permissions get destroyed alongside the profile).

Storage is handled via a PrefService, which writes to a
permissions.json file stored inside the profile folder. This is
different to Chromium's implementation, which is massively
overengineered and would require enabling a ton of code we will
never need to use.

[ChangeLog][QtWebEngineCore][QWebEngineProfile] Added new API
to control permission persistence.
[ChangeLog][QtWebEngineQuick] Added new API
to control permission persistence.

Fixes: QTBUG-55108
Change-Id: Ib3057feda3bfbbf2a17a86356feca35a67180806
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>