Add zoomFactorChanged signal to QWebEnginePage and QQuickWebEngineView
Added a signal that fires whenever the zoom factor for a given page
changes, including both via key combinations, and when using the
existing zoom API.

This already existed in Quick, but was both undocumented and only
half-implemented (zooming via keyboard/mouse wouldn't trigger it).
Thus, the Quick documentation was changed to list it as a 6.8 addition.

Fixes: QTBUG-111041
Change-Id: I91eb7c9bf0d17d99956784841c5c86640554ad1c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>