PDF multipage example: don't assume http by default
Avoid doing network operations unless a network-specific scheme is
given.

In fact, network loading is not supported anyway, so
QUrl::fromLocalFile() would be just as good for now. But perhaps we
could support network loading eventually; and this approach is not
harmful in the example in the meantime. You just get a "file not found"
error if you give an http(s) URL on the command line.

Pick-to: 6.5 6.7
Fixes: QTBUG-123548
Change-Id: I2364fd99396d4d65df92106e45818166ef2abf2f
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>