Fix crash when playing video with disabled NativePixmap support
The VideoRenderer tries to create VideoFrames backed by native
textures if possible. This leads to a crash or assert if GBM and
NativePixmaps are not supported. Replace the corresponding asserts with
early returns. The renderer is able to handle the case when
NativePixmaps couldn't been created.

Also disable native texture backed VideoFrames for Nvidia by default to
avoid this code path.

Pick-to: 6.7
Change-Id: I5f271af99875dfbc28350c6d76d3bf7a1a697b9f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>