Properly release COM interfaces to silence Direct3D11 warnings
When calling QueryInterface() on a COM object, the returned resource
is actually reference counted, and needs to be properly disposed of.
This commit fixes two instances where these resources were not getting
released, which removes a bunch of Direct3D11 live object warnings
that would appear on program shutdown.

Task-number: QTBUG-123790
Change-Id: I7115215ece5fe66d3dc7c192e2156c7070ecf10f
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>