Change qt.pdf.links log category to qt.pdf.wigets.links in pdfwidgets
We use the qt.pdf.links logging category in the PDF library.
If we redefine qLcLink in PdfWidgets, we get a static link error;
if we redeclare it, we get a dynamic link error. So it seems
that we cannot reuse logging categories across modules.

Fixes: QTBUG-124506
Pick-to: 6.5
Change-Id: I591f2e3af2162274c8092397cadf74ddbf687e9d
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit dde69e62978f136ad789f44009c6d959a9289224)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>