Prevent the white flashes during loading
It was recommended to set background color based on the web content to
prevent the white flashes that may appear during load. However, it was
not working, as the render widget host view may get created while load
a new URL depending upon the PageTransition parameter value. Applied
background color to the newly created the render widget host view to
resolve the issue. This should be done after the frame is bound to the
view, else might lead to a DCHECK failure when dealing with transparent
window

Fixes: QTBUG-112013
Pick-to: 6.5 6.6 6.7
Change-Id: I200d3838b1c0628c2950e05e717215f149c11b6a
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>