Report an error if system icu is enabled outside Linux and Android
The default value for the GN argument icu_use_data_file is true. If it
is true, it will try to load the external ICU data, and if the external
data file is not present, it will result in a runtime error. system icu
option only valid for linux and android. Added code to report configure
failure when using QT_FEATURE_webengine_system_icu=ON for unsupported
platforms

Fixes: QTBUG-125452
Pick-to: 6.8 6.7 6.6
Change-Id: I4db6edbd6e45fa216df27843084e21f993492f0e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>