Ensure Chromium features are propagated to child processes
With the 118 update, the fallback code path that passed Chromium
feature flags via the command line was completely removed in favor
of a shared memory model. Correctly initializing this requires
having a FieldTrialList instance alive somewhere, but some fixups
when integrating 118 rendered that code path unreachable. As a result,
Chromium features were never getting sent to the render process,
and were only visible on the main WebEngine process.

This change adds a FieldTrialList instance inside WebEngineContext,
which allows the FeaturesList class to work correctly, and thus
correctly propagate features to the render process.

Pick-to: 6.8 6.7
Change-Id: I6d07ea4a7c87db83e526ea62ca4d86526463ad1a
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>