Fix crash on Windows when storing blobs on disk
As a security feature, files that may get accessed by multiple threads
on Windows get marked with a deny execute ACE when written to disk,
which prevents them from being opened for execution later on.
Unfortunately, the directories allowed to have files marked like this
are limited, and need explicit opt-in. This triggers a debug crash
whenever a blob is to be written to disk, since the target directory is
not in that list.

This change ensures that profile storage directories (including custom
ones set by the user) get added to that list. We do so by keeping our
own static list of all storage paths needed by non-OTR profiles, and
making sure that said list is passed to Chromium whenever a new storage
path is to be used.

However, the code in this commit only ensures that the last
registered profile path will work correctly, since Chromium does not
expect an embedder to allow multiple non-OTR profiles alive at the same
time, each with its own arbitrary storage path. To make this use case
work, the underlying code in the path service is also changed in a
separate Chromium change.

Pick-to: 6.7
Change-Id: Ic0f12f8e444fc0c6325cc2f1ee06a0b73634e398
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit d06d875cdee7cff699b0c3163e2f1a36f3158670)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>