Use device pixel ratio of the window rather than the screen
Window device pixel ratio can differ from the screen depending on the
platform. The window is the more authorative source with new events in
Qt to support that.

In Chromium the rendering is still based off a screen factor, but each
view will have a completely independent set of screen variables, so we
can adjust the screen to match our current view.

Chromium itself also alters the screen scale factor on a per window
basis for their own native Wayland backend or if recording a window.

Fixes: QTBUG-113574
Pick-to: 6.8
Change-Id: I12231c826525a9d7b5e30eed10324f65fbf8c16f
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>