Do not override signals in qml delegates
Since 36478e25 in qtdeclarative overriding signals
is reported as warning and will be dropped in the
future as it is not supported.

Drop 'accepted' and 'rejected' signals and use signals from
base component, which is the "correct" way of handling it.
Remove some logic coming from qqc1.

Change-Id: I3eeeca4b25fb866d0c6e3da01c1513afd85904f0
Reviewed-by: Michael Brüning <michael.bruning@qt.io>
(cherry picked from commit f600a8ee31ccd01739517b146f010173768e9252)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>