QWebEnginePermission: fix the name of comparison helper
Do not use comparesEqual() as a name of a 1-arg member function, as
it is used in the C++20 comparison framework and should be a 2-arg
hidden friend.
Rename the exported member function to equals(), thus leaving the
possibility to use the comparison helper macros later, if needed.

Amends 00d7de830a43a84a431511a55525f04c21b8ba49.

Found in Qt 6.8 API review.

Pick-to: 6.8
Change-Id: I0ed9bf9ee08418da2f489ed328aecee412486fd0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>