Write permissions and preferences to disk whenever they're modified
Previously, permissions and user prefs would only get committed to disk
when the application closed, and only if it did so with no errors.
This could obviously lead to data loss when the application crashes,
so this change adds calls to PrefService::SchedulePendingLossyWrites()
whenever a permission or user pref is modified. This way, the data
will be written to disk quickly after it's modified inside the
PrefService.

Pick-to: 6.8
Change-Id: I8aa7c22f9a8b701f697f39132bf3e82ae3c66cc0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>