Stop location provider when location is disabled globally
Qt Core still doesn't have a permission backend on Windows, which
results in the permission check for location always returning true,
regardless of the actual Windows setting. In that case, the WebEngine
location provider would just pass along the info that access was denied,
but keep updates on, resulting in the location request just timing out.

This change makes sure that the location provider is stopped immediately
when access is denied, removing the timeout and signalling
to the we application that permission has been denied. This isn't the
exact behavior of other browsers, since they fall back on consentless
(low accuracy) location info and we don't, but that's a limitation of
the QtPositioning module.

Pick-to: 6.7
Change-Id: I097bade1ec6b31352321139195b630124657c23a
Reviewed-by: Michael Brüning <michael.bruning@qt.io>
(cherry picked from commit ba1d8dc18f6b2406f39968ec1070ea5becfb6a8e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>