Printing: Unify PDF generation among print() and printToPdf()
For the rasterized print() job, we have been generating the PDF data
without margins and then QPrinter applied its margins when drawing to
the device. Margins were the only differences between PDF generations
for print() and printToPdf().

By taking the margins into account at generation time, and setting
QPrinter margins to null and then restoring it after the job, we can get
rid of this terrible design choice and having a unified PDF generator
backend. This results a much more simpler code and opens the way for
other developments.

Change-Id: If2dd8932073f5112e2e29d52db624db496f59684
Reviewed-by: Michael Brüning <michael.bruning@qt.io>