Avoid crash on WebEngineView destruction with ASAN enabled
The chromium's raw_ptr implementation sets the pointer to null on
destruction if ASAN is enabled. This will lead to crash when trying to
access m_factory which is null from the NativeSkiaOutputDevice::Buffer
destructor. Both buffer and m_factory are members of the class
NativeSkiaOutputDevice. Member variables are deleted in bottom-to-top
order based on their declaration. So, by the time the buffer
destructor is invoked, m_factory is set to null. Used C++ raw pointer
instead of chromium’s raw_ptr and added a nullptr check.

Change-Id: Idfad2f5b1bb2adf8923a8fab872fdbaedf6c49f9
Pick-to: 6.7
Fixes: QTBUG-127611
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
(cherry picked from commit 9bcb567b6799bffe90b6d9f90da42fc331684947)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>