Do not set Vulkan rendering if RHI is disabled
If Qt is built without OpenGL support, the default graphics API is
Vulkan on Linux. In this case we initialize Chromium with Vulkan
rendering. This doesn't make too much sense if the user requests
software rendering what also disables RHI.

Try to avoid this configuration even if it doesn't cause any known issue.

Pick-to: 6.8
Change-Id: Ia70366a4edfd7204940fd17dea8763a9aa8e4747
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>